Use user data from context
fix

+++ backend/src/modules/habits/habit.controllers.ts
@@ -1,6 +1,6 @@
 import { createDb } from '@/db';
 import { habits, insertHabitSchema, patchHabitSchema } from '@/db/schema';
-import { eq } from 'drizzle-orm';
+import { and, eq } from 'drizzle-orm';
 import { Context } from 'hono';
 import * as HttpStatusCodes from 'stoker/http-status-codes';
 
@@ -11,7 +11,8 @@ export async function create(c: Context) {
   let parseResult = insertHabitSchema.safeParse(body);
   if (!parseResult.success) return c.json({ error: parseResult.error.format() }, HttpStatusCodes.UNPROCESSABLE_ENTITY);
 
-  let { name, userId, description } = parseResult.data;
+  let userId = c.get('currentUser').id;
+  let { name, description } = parseResult.data;
 
   try {
     let [newHabit] = await db.insert(habits)
@@ -26,12 +27,10 @@ export async function create(c: Context) {
 
 export async function index(c: Context) {
   let db = createDb(c.env);
-
-  let url = new URL(c.req.url);
-  let userId = url.searchParams.get('user_id');
+  let userId = c.get('currentUser').id;
 
   try {
-    let whereClause = userId ? eq(habits.userId, Number(userId)) : undefined;
+    let whereClause = userId ? eq(habits.userId, userId) : undefined;
     let allHabits = await db.select().from(habits)
       .where(whereClause)
       .all();
@@ -71,7 +70,9 @@ export async function update(c: Context) {
   let parseResult = patchHabitSchema.safeParse(body);
   if (!parseResult.success) return c.json({ error: parseResult.error.format() }, HttpStatusCodes.UNPROCESSABLE_ENTITY);
 
-  let { name, userId, description } = parseResult.data;
+  let userId = c.get('currentUser').id;
+
+  let { name, description } = parseResult.data;
 
   let updateFields = Object.fromEntries(
     Object.entries({ name, userId, description }).filter(([_, v]) => v !== undefined)
@@ -82,7 +83,7 @@ export async function update(c: Context) {
   try {
     let [updatedHabit] = await db.update(habits)
       .set(updateFields)
-      .where(eq(habits.id, id))
+      .where(and(eq(habits.id, id), eq(habits.userId, userId)))
       .returning();
 
     if (!updatedHabit) return c.json({ error: 'Habit not found' }, HttpStatusCodes.NOT_FOUND);
@@ -99,9 +100,11 @@ export async function destroy(c: Context) {
   let id = Number(c.req.param('id'));
   if (!id || Number.isNaN(id)) return c.json({ error: 'Invalid or missing habit ID' }, HttpStatusCodes.BAD_REQUEST);
 
+  let userId = c.get('currentUser').id;
+
   try {
     let result = await db.delete(habits)
-      .where(eq(habits.id, id))
+      .where(and(eq(habits.id, id), eq(habits.userId, userId)))
       .run();
     if (result.rowsAffected === 0) return c.json({ error: 'Habit not found' }, HttpStatusCodes.NOT_FOUND);
